+++ AffindaAPI/AffindaAPI/AffindaAPIModelFactory.cs
@@ -140,6 +140,7 @@ namespace Affinda.API.Models
         /// <param name="extractor"></param>
         /// <param name="autoValidationThreshold"></param>
         /// <param name="fields"></param>
+        /// <param name="fieldsLayout"></param>
         /// <param name="fieldsConfigured"></param>
         /// <param name="dateFormatPreference"></param>
         /// <param name="dateFormatFromDocument"> Predict the date format from any dates in the document that is not ambiguous. </param>
@@ -148,11 +149,11 @@ namespace Affinda.API.Models
         /// <param name="confirmedDocsCount"> Number of validated documents in the collection. </param>
         /// <param name="ingestEmail"> When you send email to this address, any document attached in the body will be uploaded to this collection. </param>
         /// <returns> A new <see cref="Models.Collection"/> instance for mocking. </returns>
-        public static Collection Collection(string identifier = null, string name = null, CollectionWorkspace workspace = null, Extractor extractor = null, float? autoValidationThreshold = null, IEnumerable<FieldGroup> fields = null, bool? fieldsConfigured = null, CollectionDateFormatPreference? dateFormatPreference = null, bool? dateFormatFromDocument = null, ExtractorConfig extractorConfig = null, int? unvalidatedDocsCount = null, int? confirmedDocsCount = null, string ingestEmail = null)
+        public static Collection Collection(string identifier = null, string name = null, CollectionWorkspace workspace = null, Extractor extractor = null, float? autoValidationThreshold = null, IEnumerable<FieldGroup> fields = null, FieldsLayout fieldsLayout = null, bool? fieldsConfigured = null, CollectionDateFormatPreference? dateFormatPreference = null, bool? dateFormatFromDocument = null, ExtractorConfig extractorConfig = null, int? unvalidatedDocsCount = null, int? confirmedDocsCount = null, string ingestEmail = null)
         {
             fields ??= new List<FieldGroup>();
 
-            return new Collection(identifier, name, workspace, extractor, autoValidationThreshold, fields?.ToList(), fieldsConfigured, dateFormatPreference, dateFormatFromDocument, extractorConfig, unvalidatedDocsCount, confirmedDocsCount, ingestEmail);
+            return new Collection(identifier, name, workspace, extractor, autoValidationThreshold, fields?.ToList(), fieldsLayout, fieldsConfigured, dateFormatPreference, dateFormatFromDocument, extractorConfig, unvalidatedDocsCount, confirmedDocsCount, ingestEmail);
         }
 
         /// <summary> Initializes a new instance of CollectionWorkspace. </summary>
@@ -195,6 +196,44 @@ namespace Affinda.API.Models
         public static ExtractorBaseExtractor ExtractorBaseExtractor(string identifier = null, string name = null, string namePlural = null, bool validatable = default, bool? isCustom = null, DateTimeOffset? createdDt = null)
         {
             return new ExtractorBaseExtractor(identifier, name, namePlural, validatable, isCustom, createdDt);
+        }
+
+        /// <summary> Initializes a new instance of FieldsLayout. </summary>
+        /// <param name="defaultCategory"></param>
+        /// <param name="categories"></param>
+        /// <returns> A new <see cref="Models.FieldsLayout"/> instance for mocking. </returns>
+        public static FieldsLayout FieldsLayout(FieldCategory defaultCategory = null, IEnumerable<FieldCategory> categories = null)
+        {
+            categories ??= new List<FieldCategory>();
+
+            return new FieldsLayout(defaultCategory, categories?.ToList());
+        }
+
+        /// <summary> Initializes a new instance of FieldCategory. </summary>
+        /// <param name="enabledFields"></param>
+        /// <param name="disabledFields"></param>
+        /// <returns> A new <see cref="Models.FieldCategory"/> instance for mocking. </returns>
+        public static FieldCategory FieldCategory(IEnumerable<Field> enabledFields = null, IEnumerable<Field> disabledFields = null)
+        {
+            enabledFields ??= new List<Field>();
+            disabledFields ??= new List<Field>();
+
+            return new FieldCategory(enabledFields?.ToList(), disabledFields?.ToList());
+        }
+
+        /// <summary> Initializes a new instance of Field. </summary>
+        /// <param name="label"></param>
+        /// <param name="dataPoint"> Data point identifier. </param>
+        /// <param name="mandatory"></param>
+        /// <param name="autoValidationThreshold"></param>
+        /// <param name="showDropdown"></param>
+        /// <param name="fields"></param>
+        /// <returns> A new <see cref="Models.Field"/> instance for mocking. </returns>
+        public static Field Field(string label = null, string dataPoint = null, bool? mandatory = null, float? autoValidationThreshold = null, bool? showDropdown = null, IEnumerable<Field> fields = null)
+        {
+            fields ??= new List<Field>();
+
+            return new Field(label, dataPoint, mandatory, autoValidationThreshold, showDropdown, fields?.ToList());
         }
 
         /// <summary> Initializes a new instance of PathsOxm5M7V3DocumentsGetResponses200ContentApplicationJsonSchema. </summary>