+++ AffindaAPI/AffindaAPI/Models/Field.Serialization.cs
@@ -11,73 +11,21 @@ using Azure.Core;
 
 namespace Affinda.API.Models
 {
-    public partial class Field : IUtf8JsonSerializable
+    public partial class Field
     {
-        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WritePropertyName("label");
-            writer.WriteStringValue(Label);
-            if (Optional.IsDefined(Slug))
-            {
-                writer.WritePropertyName("slug");
-                writer.WriteStringValue(Slug);
-            }
-            writer.WritePropertyName("dataPoint");
-            writer.WriteStringValue(DataPoint);
-            if (Optional.IsDefined(Mandatory))
-            {
-                writer.WritePropertyName("mandatory");
-                writer.WriteBooleanValue(Mandatory.Value);
-            }
-            if (Optional.IsDefined(Disabled))
-            {
-                writer.WritePropertyName("disabled");
-                writer.WriteBooleanValue(Disabled.Value);
-            }
-            if (Optional.IsDefined(AutoValidationThreshold))
-            {
-                writer.WritePropertyName("autoValidationThreshold");
-                writer.WriteNumberValue(AutoValidationThreshold.Value);
-            }
-            if (Optional.IsDefined(ShowDropdown))
-            {
-                writer.WritePropertyName("showDropdown");
-                writer.WriteBooleanValue(ShowDropdown.Value);
-            }
-            if (Optional.IsCollectionDefined(Fields))
-            {
-                writer.WritePropertyName("fields");
-                writer.WriteStartArray();
-                foreach (var item in Fields)
-                {
-                    writer.WriteObjectValue(item);
-                }
-                writer.WriteEndArray();
-            }
-            writer.WriteEndObject();
-        }
-
         internal static Field DeserializeField(JsonElement element)
         {
             string label = default;
-            Optional<string> slug = default;
             string dataPoint = default;
             Optional<bool> mandatory = default;
-            Optional<bool> disabled = default;
-            Optional<float> autoValidationThreshold = default;
+            Optional<float?> autoValidationThreshold = default;
             Optional<bool> showDropdown = default;
-            Optional<IList<Field>> fields = default;
+            Optional<IReadOnlyList<Field>> fields = default;
             foreach (var property in element.EnumerateObject())
             {
                 if (property.NameEquals("label"))
                 {
                     label = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("slug"))
-                {
-                    slug = property.Value.GetString();
                     continue;
                 }
                 if (property.NameEquals("dataPoint"))
@@ -95,21 +43,11 @@ namespace Affinda.API.Models
                     mandatory = property.Value.GetBoolean();
                     continue;
                 }
-                if (property.NameEquals("disabled"))
-                {
-                    if (property.Value.ValueKind == JsonValueKind.Null)
-                    {
-                        property.ThrowNonNullablePropertyIsNull();
-                        continue;
-                    }
-                    disabled = property.Value.GetBoolean();
-                    continue;
-                }
                 if (property.NameEquals("autoValidationThreshold"))
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
-                        property.ThrowNonNullablePropertyIsNull();
+                        autoValidationThreshold = null;
                         continue;
                     }
                     autoValidationThreshold = property.Value.GetSingle();
@@ -141,7 +79,7 @@ namespace Affinda.API.Models
                     continue;
                 }
             }
-            return new Field(label, slug.Value, dataPoint, Optional.ToNullable(mandatory), Optional.ToNullable(disabled), Optional.ToNullable(autoValidationThreshold), Optional.ToNullable(showDropdown), Optional.ToList(fields));
+            return new Field(label, dataPoint, Optional.ToNullable(mandatory), Optional.ToNullable(autoValidationThreshold), Optional.ToNullable(showDropdown), Optional.ToList(fields));
         }
     }
 }